Showing book list status title, small fixes

+++ app/src/main/java/com/littletrickster/usersbooks/db/models/Book.kt
@@ -84,7 +84,7 @@ interface BookDao {
     WHERE b.id = :id
     """
     )
-    fun getBookWithTitleById(id: Long): Flow<BookWithListTitle?>
+    fun getBookWithTitleById(id: Int): Flow<BookWithListTitle?>
